+++ src/measure-blocks/image.ts
@@ -8,7 +8,7 @@ import type {
 } from '../types-internal.js'
 import { PretextPdfError } from '../errors.js'
 import { measureRichText } from '../rich-text.js'
-import { buildFontKey } from '../measure.js'
+import { buildFontKey } from '../font-key.js'
 import { measureText } from '../measure-text.js'
 import { LINE_HEIGHT_BODY } from '../render-utils.js'