+++ src/render-utils.ts
@@ -5,11 +5,7 @@
 
 import { PDFDocument, PDFFont, PDFName, PDFNull, PDFRef, PDFString, rgb } from '@cantoo/pdf-lib'
 import { PretextPdfError } from './errors.js'
-
-// NOTE: this pattern is duplicated in src/validate/helpers.ts:SAFE_URL_SCHEME.
-// Importing from validate/ in a render module creates a wrong-direction
-// dependency. Consolidate to a shared src/url-utils.ts leaf module in v1.8.
-const SAFE_URL_SCHEME = /^(https?|mailto|ftp|#)/i
+import { SAFE_URL_SCHEME } from './url-utils.js'
 
 /**
  * Draw a single line of text with justified alignment.